fix: remove duplicate implementations and add missing helper functions
- Remove refund.rs and release.rs (duplicate implementations)
- Add missing helper functions: is_initialized, get_protocol_fee_bps, calculate_protocol_fee
- Add AmountMustBePositive error code
- Import contracterror macro
- Fix duplicate id variable in create_contract.rs
- Fix InvalidParticipant vs InvalidParticipants naming

Remaining: EscrowClient/EscrowArgs import issues in module files (8 errors)
